requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
adwaita-icon-theme=41.0=ha770c72_1
atk-1.0=2.36.0=h7fe96df_0
bottleneck=1.3.4=py39heec5a64_0
c-ares=1.18.1=h3422bc3_0
ca-certificates=2022.3.29=hca03da5_0
cairo=1.16.0=h0ab239c_1
camelot-py=0.9.0=pyhd8ed1ab_0
certifi=2021.10.8=py39hca03da5_2
cffi=1.15.0=py39h22df2f2_1
chardet=4.0.0=py39h2804cbe_1
click=7.1.2=pyh9f0ad1d_0
cryptography=35.0.0=py39hfb8cd70_1
cycler=0.11.0=pyhd8ed1ab_0
dbus=1.13.18=h8280c03_0
eigen=3.4.0=hc021e02_0
epoxy=1.5.10=h1c322ee_1
et_xmlfile=1.0.1=py_1001
expat=2.4.4=hc377ac9_0
fontconfig=2.13.94=heb65262_0
freetype=2.10.4=h17b34a0_1
fribidi=1.0.10=h27ca646_0
gdk-pixbuf=2.42.6=h74ef11a_4
gettext=0.21.0=h826f4ad_0
ghostscript=9.54.0=hbdafb3b_1
giflib=5.2.1=h27ca646_2
glib=2.69.1=h98b2900_1
gmp=6.2.1=h9f76cd9_0
gobject-introspection=1.68.0=py39h3459c9f_3
graphite2=1.3.14=h0e5e14a_0
gst-plugins-base=1.14.1=hf0a386a_0
gstreamer=1.14.1=he09cfb7_0
gtk3=3.24.21=h321fa23_1
harfbuzz=2.8.1=hffeda63_0
hdf5=1.12.1=nompi_had0e5e0_100
hicolor-icon-theme=0.17=hce30654_2
icu=68.2=hbdafb3b_0
jbig=2.1=h3422bc3_2003
jpeg=9e=h1c322ee_1
kiwisolver=1.3.2=py39hc377ac9_0
krb5=1.19.2=h3b8d789_0
lcms2=2.12=had6a04f_0
lerc=2.2.1=h9f76cd9_0
libblas=3.9.0=14_osxarm64_openblas
libcblas=3.9.0=14_osxarm64_openblas
libcurl=7.82.0=hc6d1d07_0
libcxx=12.0.0=hf6beb65_1
libdeflate=1.7=h27ca646_5
libedit=3.1.20210910=h1a28f6b_0
libev=4.33=h642e427_1
libevent=2.1.10=hbae9a57_4
libffi=3.4.2=hc377ac9_2
libgfortran=5.0.0.dev0=11_0_1_hf114ba7_23
libgfortran5=11.0.1.dev0=hf114ba7_23
libiconv=1.16=h642e427_0
liblapack=3.9.0=14_osxarm64_openblas
libnghttp2=1.46.0=h95c9599_0
libopenblas=0.3.20=openmp_h2209c59_0
libpng=1.6.37=hf7e6567_2
libpq=12.9=h65cfe13_1
librsvg=2.50.7=h90a7944_0
libssh2=1.10.0=hb80f160_2
libtiff=4.3.0=hc6122e1_1
libwebp=1.2.2=h0d20362_0
libwebp-base=1.2.2=h3422bc3_1
libxml2=2.9.12=h8c5e841_0
libxslt=1.1.34=h9833966_0
llvm-openmp=13.0.1=h455960f_1
lz4-c=1.9.3=hbdafb3b_1
matplotlib-base=3.4.3=py39h5aa4fe7_1
ncurses=6.3=h1a28f6b_2
ninja=1.10.2=hc021e02_1
nspr=4.33=hc377ac9_0
nss=3.74=h142855e_0
numexpr=2.8.1=py39h144ceef_0
numpy=1.19.5=py39h1f3b974_2
opencv=4.5.4=py39h7dfc772_3
openpyxl=3.0.9=pyhd8ed1ab_0
openssl=1.1.1n=h1a28f6b_0
packaging=21.3=pyhd8ed1ab_0
pandas=1.3.4=py39h9197a36_0
pango=1.48.7=h79d33a7_0
pcre=8.45=hbdafb3b_0
pdfminer.six=20220319=pyhd8ed1ab_0
pillow=9.0.1=py39h4d1bdd5_0
pip=21.2.4=py39hca03da5_0
pixman=0.40.0=h27ca646_0
pycparser=2.21=pyhd8ed1ab_0
pycryptodome=3.12.0=py39h695f6f0_0
pyparsing=3.0.8=pyhd8ed1ab_0
pypdf2=1.27.6=pyhd8ed1ab_0
python=3.9.12=hbdb9e5c_0
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.9=2_cp39
pytz=2022.1=pyhd8ed1ab_0
qt=5.15.2=h4e944ae_2
readline=8.1.2=h1a28f6b_1
setuptools=61.2.0=py39hca03da5_0
six=1.16.0=pyh6c4a22f_0
sqlite=3.38.2=h1058600_0
tk=8.6.11=hb8d0fd4_0
tornado=6.1=py39h5161555_1
tzdata=2022a=hda174b7_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.5=h1a28f6b_0
zlib=1.2.12=h5a0b063_1
zstd=1.5.0=h8574219_0
